+++ Modules/Officer.py
@@ -4,6 +4,7 @@ import Constants
 from Logger import logger
 from discord.utils import get
 import discord, asyncio, re, emoji, csv
+from discord.errors import NotFound
 
 class Officer(commands.Cog):
     def __init__(self, bot):
@@ -160,11 +161,19 @@ class Officer(commands.Cog):
     async def send_intro_qn(self, ctx, msg):
         if is_brioche_bun(ctx.message.author.roles):
             about_us = self.bot.get_channel(Constants.ID_CHN_ABOUT_US)
-            member = await self.bot.fetch_user(msg[3:-1])
-            logger.info("{0} invoked intro for {1}".format(ctx.message.author.display_name, member.display_name))
-            await ctx.message.delete()
-            await ctx.send(file=discord.File(Constants.ASSET_POSTER))
-            await ctx.send(Constants.MSG_REC_OPEN_MSG.format(member, about_us))
+            try:
+                member = await self.bot.fetch_user(msg[3:-1])
+                logger.info("{0} invoked intro for {1}".format(ctx.message.author.display_name, member.display_name))
+                await ctx.send(Constants.MSG_REC_OPEN_MSG.format(member.mention, about_us))
+            except NotFound:
+                logger.info("Unknown user fetched by {0}, setting member to empty string and sending generic welcome message instead".format(ctx.message.author.display_name))
+                member = None
+            finally:
+                await ctx.message.delete()
+                await ctx.send(file=discord.File(Constants.ASSET_POSTER))
+                await ctx.send(Constants.MSG_REC_OPEN_MSG.format(member.mention, about_us)) \
+                if member else await ctx.send(Constants.MSG_REC_OPEN_MSG.format('', about_us))
+
         else:
             logger.warn("Member %s tried to call intro", ctx.message.author.display_name)